# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PCPartPicker API
  version: 1.0.0
  description: |
    REST API for the PCPartPicker prototype. The client must never access the Postgres/Supabase
    database directly; all reads and writes go through this API. Each resource exposes standard
    CRUD operations (create, read, update, delete) for the project’s tables.
servers:
  - url: http://localhost:3000
    description: Local dev server
paths:
...
  /user-profiles:
    get:
      summary: List user profiles
      tags: [UserProfiles]
      responses:
        '200':
          description: List of user profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
    post:
      summary: Create a user profile
      description: Creates an application-level profile for an existing auth user.
      tags: [UserProfiles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileCreate'
      responses:
        '201':
          description: Created profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /user-profiles/{userId}:
    get:
      summary: Get a user profile by userId
      tags: [UserProfiles]
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: Profile not found
    patch:
      summary: Update a user profile
      tags: [UserProfiles]
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    delete:
      summary: Delete a user profile
      tags: [UserProfiles]
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: Deleted

  # -------------------------
  # VEHICLES
  # -------------------------
...
  /listings:
    get:
      summary: List listings
      tags: [Listings]
      parameters:
        - in: query
          name: status
          schema:
            type: string
          description: Filter by listing status (active, sold, etc.)
        - in: query
          name: source
          schema:
            type: string
          description: Filter by source (dealer, private)
      responses:
        '200':
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Listing'
    post:
      summary: Create a listing
      tags: [Listings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCreate'
      responses:
        '201':
          description: Created listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
  /listings/{listingId}:
    get:
      summary: Get listing by ID
      tags: [Listings]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200':
          description: Listing with related vehicle info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '404':
          description: Not found
    patch:
      summary: Update listing
      tags: [Listings]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingUpdate'
      responses:
        '200':
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
    delete:
      summary: Delete listing
      tags: [Listings]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '204':
          description: Deleted
  /listings/{listingId}/photos:
    get:
      summary: List photos for a listing
      tags: [ListingPhotos]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200':
          description: Photos for listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListingPhoto'
    post:
      summary: Add a photo to a listing
      tags: [ListingPhotos]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingPhotoCreate'
      responses:
        '201':
          description: Created photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingPhoto'
  /listing-photos/{photoId}:
    delete:
      summary: Delete a listing photo
      tags: [ListingPhotos]
      parameters:
        - $ref: '#/components/parameters/PhotoIdParam'
      responses:
        '204':
          description: Deleted
...